Start migrating to Angular CLI
As of this commit the project is in a non-working state. This is a
transitional commit created because there are too many changes.

Moved the TypeScript and SCSS files to the Client folder which is
the root of the Angular CLI project. Ejected the project to use
Webpack directly. Installed the necessary node modules.