Upgrade to .NET Core 2.0
Upgraded the app to .NET Core 2.0 and instead of referencing
multiple NuGet packages it now references the Microsoft.AspNetCore.All
metapackage.

Also moved the API controllers into the Controllers folder because
it contatined only one controller - HomeController.

The routing doesn't work for some reason so the API is inaccessible
for clients right now.