Add the groundwork for schedule services
Although the schedule client and the converter into calendar classes
are not implemented yet, some groundwork for them has been done.

I created the classes and interfaces that represent the services and
options. I also created some extension methods to include in the
startup service configuration pipeline and options to store in
appsettings.json or configure in code.