Make HttpClient a dependency of ScheduleHttpClient
Creating a new instance of HttpClient directly in ScheduleHttpClient
is a bad idea, that's why I added it to the DI container to handle.
Although, the DI container just creates a new instance for every
scope which is exactly what the ScheduleHttpClient did before, but
I guess it's more flexible this way.