Change the name and structure of the project
This is a massive commit which includes a complete overhaul of the
project. It's not recommended to revert back before this point.

Changed the name to Cloud Calendar, because Interlogic Project was
just a working title.

Also created a new project in Visual Studio using the ASP.NET Core/
Angular template whick includes Webpack and server-side rendering
by default. Copied all the previous project's files into the new
project (including the .git folder).

Here are the following major changes to the project:

 - Client-side code now resides in the ClientApp folder.

 - SCSS files are moved to the css folder.

 - Deleted the footer because of some CSS problems which I'll
investigate later.

 - Disabled the default server-side rendering, because jQuery doesn't
support running on the server and I cannot get rid of it, because
I use FullCalendar.js which is an essential part of the app, and it's
a jQuery plugin. All of the server-side enabling code is still in
place, though.

 - The webpack configuration is now the default one from the template
(with minimal tweaks).

 - The Index.cshtml view now contains only the essential parts.
Everything else was moved into a _Layout.

 - Added a dummy error handler for Production environment.

 - The projects are no longer in the src and test folders.

 - There are a lot of other changes, but they are quite small.