Test and fix CreateCalendar in CalendarService
Fixed the CreateCalendar with the help of a unit test.

Also, added schedule - the way to get the start and end time of a
class from its number, and class duration. Added the schedule to
appsettings. Configured the service pipeline to add schedule info.

Moved all the options from the Schedule project into Services.Options
folder.

Also, updated packages.