Create an API endpoint for configuration
Created the API controller for schedule options and app settings.
Also created a service on the front-end to fetch the data.

Now the class duration is not hard-coded at 1 hour 20 minutes, but
provided in the configuration which is fetched from the server.

Also updated the documentation of the API controllers.