Start implementing admin pages
Started implementing the settings and users pages.

The settings page is for creating/editing/deleting faculties,
groups, classrooms, buildings and subjects. Only groups are yet
to be implemented.

The users page is for creating/editing/deleting students and
lecturers. It is not implemented yet.

Also, removed the faculty-building reference as it is redundant and
comlicates creating faculties.